requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-docx>=0.8.11
pandas>=2.0.0
openpyxl>=3.1.0
selenium>=4.15.0
webdriver-manager>=4.0.0 